Worst commit ever, but we can now save data from the button on the GUI inside "Opret Sag", currently the data generation is hardcoded as i could not find an easy way to get the values from the table, we will have to look into that, this should not be pushed to master at the current state.
As the Facades havn't been setup yet i had to change a few cunstructors and import a few classes directly in the presentation lay that should be handled by some interfaces or facades, same goes for buissiness to data.

On a brighter note: Fixed a few issues with the PostgresHelperClass that were hard to track down, as they didn't present any errors in any form but made the sql not run.